% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFactors.R
\docType{data}
\name{diffFactors}
\alias{diffFactors}
\title{diffFactors}
\format{
A data frame with 483 rows and 32 variables
}
\usage{
data(diffFactors)
}
\description{
Data for the comparison of COVID-19 mortality in European and North American geographic entities
}
\details{
\itemize{
\item geographic_entity_receptor are the entity receptor
\item geographic_entity_source are the entity source
  \item proba is the probability that the receptor follows the mortality dynamics of the source
  \item other columns are the difference between factors
}
}
\references{
\doi{10.5281/zenodo.4769671}
}
\author{
Melina Ribaud, Davide Martinetti and Samuel Soubeyrand
}
\keyword{diffFactors}
