% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Zero-inflated proportions dataset}
\format{
A data frame with 440 rows and 23 variables
}
\usage{
data(example_data)
}
\description{
A dataset example to test the package functions.
The factor X1 to X5 and F1 to F5 are correlated to the responses y.
}
\details{
\itemize{
\item ID.source are the ID of source hosts
\item ID.recep are the ID of receiver hosts
  \item y are the vector of transmission probabilities source -> receiver
  \item X1 to X10 are continuous factor
  \item F1 to F10 are discrete factor
}
}
\keyword{example_data}
