% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact2mat.R
\name{fact2mat}
\alias{fact2mat}
\title{Turn factor into multiple column}
\usage{
fact2mat(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
Columns with zeros and ones.
}
\description{
Turns a factor with several levels into a matrix with several columns composed of zeros and ones.
}
\examples{
x = sample(1:3,100,replace = TRUE)
fact2mat(x)
}
