% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{The performance indicator}
\usage{
indicator(X, proba)
}
\arguments{
\item{X}{a vector.}

\item{proba}{a zero-inflated proportions response.}
}
\value{
a scalar represents the performance indicator
and the vector \code{proba}.
}
\description{
Calculate the indicator for a vector \code{X}
and a zero-inflated proportions response \code{proba}.
}
\examples{
X = rnorm(100)
proba = runif(100)
proba[sample(1:100,80)]=0
print(indicator(X,proba))
}
