% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model_matrix}
\alias{model_matrix}
\title{Construct Design Matrix}
\usage{
model_matrix(DT, ColNameFactor, other_class)
}
\arguments{
\item{DT}{a data table contains the factors and the response.}

\item{ColNameFactor}{a char vector with the name of the selected factor.}

\item{other_class}{a char vector with the name of other classes than numeric (factor or char).}
}
\value{
return the value.
}
\description{
Creates a design matrix
by expanding factors to a set of dummy variables.
}
\examples{
library(data.table)
data(example_data)
m = model_matrix (example_data,
colnames(example_data)[-c(1:3)],
other_class = colnames(example_data)[14:23])
print(m)
}
