% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_01.R
\name{scale_01}
\alias{scale_01}
\title{Scale vector}
\usage{
scale_01(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
the scaled vector of \code{x}.
}
\description{
Scale a vector between 0 and 1.
}
\examples{
x = runif(100,-10,10)
x_scale = scale_01(x)
range(x_scale)
}
