% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_stat_discr.R
\name{T_stat_discr}
\alias{T_stat_discr}
\title{Statistic for non-numeric factor tests}
\usage{
T_stat_discr(permu, al)
}
\arguments{
\item{permu}{the response vector.}

\item{al}{the factor.}
}
\value{
the statistic.
}
\description{
Statistic for non-numeric factor tests (same statistic as H-test).
}
\examples{
permu = runif(100,-10,10)
al = as.factor(sample(1:3,100,replace=TRUE))
T_stat_discr(permu, al)
}
