% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_stat_multi.R
\name{T_stat_multi}
\alias{T_stat_multi}
\title{Statistic for non-numeric factor multiple tests}
\usage{
T_stat_multi(permu, al)
}
\arguments{
\item{permu}{the response vector.}

\item{al}{the factor.}
}
\value{
the means difference of two levels for a discrete factor.
}
\description{
Statistic for non-numeric factor multiple tests (difference in mean ranks).
}
\examples{
permu = runif(100,-10,10)
al = as.factor(sample(1:3,100,replace=TRUE))
T_stat_multi(permu, al)
}
